helm: add control plane helm chart
adds a helm chart that deploys our control plane nodes (along with a
load balancer). note that the load balancer does not work out of the box
since the cluster requires that the metallb helm chart is installed on
the cluster. soon this will be automated as well with ArgoCD, though